Cart class logic done

+++ src/com/comp/Shippable.java
@@ -2,5 +2,5 @@ package com.comp;
 
 public interface Shippable {
     String getName();
-    float getWeight();
+    double getWeight();
 }
